+++ Package/Core/Promises/Internal/Continuations/CallbackHelperInternal.cs
