+++ Package/Core/Promises/Internal/Continuations/CatchPromiseInternal.cs
