+++ Package/Core/Promises/Internal/Continuations/ContinuePromiseInternal.cs
