+++ Package/Core/Promises/Internal/Continuations/PromiseSynchronousWaiterInternal.cs
