+++ Package/Core/Promises/Internal/Continuations/ThenPromiseInternal.cs
