+++ Package/Core/Promises/Internal/Groups/MergeInternal.cs
