+++ Package/Core/Promises/Internal/Groups/RaceInternal.cs
