--- Package/Core/Promises/Internal/InterfacesInternal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 49b6252da48564e49899c775e41ecdac
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
